server/internal/chunks: remove chunks package
The new chunksums API has obsoleted the need for the chunks package. The
few utilities that were useful have been swept into the packages most
appropriate for their use, and the package has been removed.

Also, remove the oppbench command, which required updates to accommodate
these changes. It was only used prove Ahmdal's law to some colleagues.
They are sold now, so it can go.
